+++ docs/changelog.md
@@ -2,6 +2,12 @@
 
 ## Changelog
 
+- v1.3.0-alpha
+  - Added mouse events (buggy)
+  - Added display._get_canvas_element()
+  - Added rect.collidepoint()
+  - 2nd June 2023
+
 - v1.2.0-alpha
   - Added rect properties (experimental)
   - 2nd June 2023